# Conflicts:
#	src/main/resources/application.yaml

+++ src/main/java/com/clearticket/clearticket/repository/TagRepository.java
@@ -3,5 +3,5 @@ package com.clearticket.clearticket.repository;
 import com.clearticket.clearticket.model.entity.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TagRepository extends JpaRepository<Tag, Long> {
+public interface TagRepository extends JpaRepository<Tag, Integer> {
 }
